Move common to a private module
