+++ src/scikit_build_core/file_api/model/toolchains.py
@@ -2,7 +2,7 @@ import dataclasses
 from pathlib import Path
 from typing import List, Optional
 
-from .common import APIVersion
+from ._common import APIVersion
 
 __all__ = ["Compiler", "Implicit", "Toolchain", "Toolchains"]
